Simplify instruction debug info decoding and cont impl of instructions
Instead of passing the opcode to our  instruction_info helper we just
pass a reference to Cpu and read instead; this gives more flexibility
obviously.

Some work has been put into adding the IX+* & IY+* modes but it's not
pretty. Some more work needs to be done in tidying this up, preferably
perhaps utilizing a common addressing mode. For now though the focus is
on passing larger parts of the Zexdoc & Zexall cpu exercisers as we're
well on our way.