Fix LD reg, IX+* addressing
This is still very much hacky but this fixes the IY+*,* / IX type
addressing modes where used in LD & MVI (the whole LD instruction
implementation could be reworked).

The goal with this commit however is to get the CPU up to a working
state; we're almost passing the Zexdoc tests now. Once passing we can
start refactoring while testing towards existing tests.

LTO is turned off currently for debugging purposes.