Fix stupid overflow flag bug & add unit tests
We had a stupid bug with the overflow flag where `overflowing_sub()` &
`overflowing_add()` was used to determine an overflow wrong.

This would only indicate whether the `u8` had overflowed it's bits, but
this does not give any information whether or not an overflow into the
sign bit has occurred...

Now understanding how the overflow flag works with regards to two's
compliment addition & subtraction, unit tests have been added (more are needed).

For addition:
`((a >> 7) == (b >> 7) && (a >> 7) != (result >> 7))`

Check bit 7 of a is equal to bit 7 of b and bit 7 of a is not equal to
the bit 7 of the result.

Example where `overflowing_add()` & the method above would work the
"same":

            0b0110_0100 (100)
          + 0b0011_0001 (49)
          = 0b1001_0101 <-- overflow has occurred
          100 + 49 = 149 (which is greater than +127).

`overflowing_add()` would not catch the following scenarios (which is
cool because this is literally an XOR truth table..):

num1 num2 sum overflow:
-----------------------
0    0    0
0    0    1   yes
0    1    0
0    1    1
1    0    0   yes
1    0    1
1    1    0
1    1    1  <-- afaik `overflowing_add()` would be true here if the
addition is greater than u8::MAX.

Likewise with subtraction an overflow can only occur if our two values
have a different sign, `overflowing_sub()` would not care about that.

The addition of simple unit tests for this hopefully stop regressions.
TODO: Need to add a larger test that tests all of the instructions that
use overflow instead of parity.